renamed package to nmealibrary
