+++ android/src/main/java/com/nmealibrary/NmeaMessageListener.java
@@ -1,4 +1,4 @@
-package com.reactlibrary;
+package com.nmealibrary;
 
 import android.location.GpsStatus.NmeaListener;
