+++ android/src/main/java/com/nmealibrary/RNNmeaLibraryModule.java
@@ -1,4 +1,4 @@
-package com.reactlibrary;
+package com.nmealibrary;
 
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
