+++ android/src/main/java/com/nmealibrary/RNNmeaLibraryPackage.java
@@ -1,5 +1,5 @@
 
-package com.reactlibrary;
+package com.nmealibrary;
 
 import java.util.Arrays;
 import java.util.Collections;
